requirements.txt
tk
customtkinter
Pillow
CTkMessagebox